Discretizes a Vector

This function takes a vector x and returns a list with information on disretized version of x, see return for details on the resulting object.

auto_cut(
  x,
  breaks = NULL,
  n_bins = 27,
  cut_type = c("equal", "quantile"),
  x_name = "value",
  level_name = "level",
  ...
)

Arguments

x
A vector.

breaks
An optional vector of breaks. Only relevant for numeric x.

n_bins
If x is numeric and no breaks are provided, this is the maximum number of bins allowed or to be created (approximately).

cut_type
For the default type "equal", bins of equal width are created by pretty. Choose "quantile" to create quantile bins.

x_name
Column name with the values of x in the output.

level_name
Column name with the bin labels of x in the output.

...
Further arguments passed to cut3.

Value

A list with the following four elements:

data A data.frame with colums x_name and level_name each with the same length as x. The column x_name has values in output bin_means while the column level_name has values in bin_labels.

breaks A vector of increasing and unique breaks used to cut a numeric x with too many distinct levels. NULL otherwise.

bin_means The midpoints of subsequent breaks, or if there are no breaks in the output, factor levels or distinct values of x.

bin_labels Break labels of the form "(low, high]" if there are breaks in the output, otherwise the same as bin_means. Same order as bin_means.

Details

The construction of level names can be controlled by passing ... arguments to formatC.

Examples

auto_cut(1:10, n_bins = 3)

#> $data
#>    value   level
#> 1    2.5  [0, 5]
#> 2    2.5  [0, 5]
#> 3    2.5  [0, 5]
#> 4    2.5  [0, 5]
#> 5    2.5  [0, 5]
#> 6    7.5 (5, 10]
#> 7    7.5 (5, 10]
#> 8    7.5 (5, 10]
#> 9    7.5 (5, 10]
#> 10   7.5 (5, 10]
#> 
#> $breaks
#> [1]  0  5 10
#> 
#> $bin_means
#> [1] 2.5 7.5
#> 
#> $bin_labels
#> [1] [0, 5]  (5, 10]
#> Levels: [0, 5] (5, 10]
#> 

auto_cut(c(NA, 1:10), n_bins = 3)

#> $data
#>    value   level
#> 1     NA    <NA>
#> 2    2.5  [0, 5]
#> 3    2.5  [0, 5]
#> 4    2.5  [0, 5]
#> 5    2.5  [0, 5]
#> 6    2.5  [0, 5]
#> 7    7.5 (5, 10]
#> 8    7.5 (5, 10]
#> 9    7.5 (5, 10]
#> 10   7.5 (5, 10]
#> 11   7.5 (5, 10]
#> 
#> $breaks
#> [1]  0  5 10
#> 
#> $bin_means
#> [1] 2.5 7.5  NA
#> 
#> $bin_labels
#> [1] [0, 5]  (5, 10] <NA>   
#> Levels: [0, 5] (5, 10]
#> 

auto_cut(1:10, breaks = 3:4, n_bins = 3)

#> $data
#>    value  level
#> 1     NA   <NA>
#> 2     NA   <NA>
#> 3    3.5 [3, 4]
#> 4    3.5 [3, 4]
#> 5     NA   <NA>
#> 6     NA   <NA>
#> 7     NA   <NA>
#> 8     NA   <NA>
#> 9     NA   <NA>
#> 10    NA   <NA>
#> 
#> $breaks
#> [1] 3 4
#> 
#> $bin_means
#> [1] 3.5  NA
#> 
#> $bin_labels
#> [1] [3, 4] <NA>  
#> Levels: [3, 4]
#> 

auto_cut(1:10, n_bins = 3, cut_type = "quantile")

#> $data
#>    value   level
#> 1    2.5  [1, 4]
#> 2    2.5  [1, 4]
#> 3    2.5  [1, 4]
#> 4    2.5  [1, 4]
#> 5    5.5  (4, 7]
#> 6    5.5  (4, 7]
#> 7    5.5  (4, 7]
#> 8    8.5 (7, 10]
#> 9    8.5 (7, 10]
#> 10   8.5 (7, 10]
#> 
#> $breaks
#> [1]  1  4  7 10
#> 
#> $bin_means
#> [1] 2.5 5.5 8.5
#> 
#> $bin_labels
#> [1] [1, 4]  (4, 7]  (7, 10]
#> Levels: [1, 4] (4, 7] (7, 10]
#> 

auto_cut(LETTERS[4:1], n_bins = 2)

#> $data
#>   value level
#> 1     D     D
#> 2     C     C
#> 3     B     B
#> 4     A     A
#> 
#> $breaks
#> NULL
#> 
#> $bin_means
#> [1] "A" "B" "C" "D"
#> 
#> $bin_labels
#> [1] "A" "B" "C" "D"
#> 

auto_cut(factor(LETTERS[1:4], LETTERS[4:1]), n_bins = 2)

#> $data
#>   value level
#> 1     A     A
#> 2     B     B
#> 3     C     C
#> 4     D     D
#> 
#> $breaks
#> NULL
#> 
#> $bin_means
#> [1] D C B A
#> Levels: D C B A
#> 
#> $bin_labels
#> [1] D C B A
#> Levels: D C B A
#> 

auto_cut(990:1100, n_bins = 3, big.mark = "'", format = "fg")

#> $data
#>     value          level
#> 1     975   [950, 1'000]
#> 2     975   [950, 1'000]
#> 3     975   [950, 1'000]
#> 4     975   [950, 1'000]
#> 5     975   [950, 1'000]
#> 6     975   [950, 1'000]
#> 7     975   [950, 1'000]
#> 8     975   [950, 1'000]
#> 9     975   [950, 1'000]
#> 10    975   [950, 1'000]
#> 11    975   [950, 1'000]
#> 12   1025 (1'000, 1'050]
#> 13   1025 (1'000, 1'050]
#> 14   1025 (1'000, 1'050]
#> 15   1025 (1'000, 1'050]
#> 16   1025 (1'000, 1'050]
#> 17   1025 (1'000, 1'050]
#> 18   1025 (1'000, 1'050]
#> 19   1025 (1'000, 1'050]
#> 20   1025 (1'000, 1'050]
#> 21   1025 (1'000, 1'050]
#> 22   1025 (1'000, 1'050]
#> 23   1025 (1'000, 1'050]
#> 24   1025 (1'000, 1'050]
#> 25   1025 (1'000, 1'050]
#> 26   1025 (1'000, 1'050]
#> 27   1025 (1'000, 1'050]
#> 28   1025 (1'000, 1'050]
#> 29   1025 (1'000, 1'050]
#> 30   1025 (1'000, 1'050]
#> 31   1025 (1'000, 1'050]
#> 32   1025 (1'000, 1'050]
#> 33   1025 (1'000, 1'050]
#> 34   1025 (1'000, 1'050]
#> 35   1025 (1'000, 1'050]
#> 36   1025 (1'000, 1'050]
#> 37   1025 (1'000, 1'050]
#> 38   1025 (1'000, 1'050]
#> 39   1025 (1'000, 1'050]
#> 40   1025 (1'000, 1'050]
#> 41   1025 (1'000, 1'050]
#> 42   1025 (1'000, 1'050]
#> 43   1025 (1'000, 1'050]
#> 44   1025 (1'000, 1'050]
#> 45   1025 (1'000, 1'050]
#> 46   1025 (1'000, 1'050]
#> 47   1025 (1'000, 1'050]
#> 48   1025 (1'000, 1'050]
#> 49   1025 (1'000, 1'050]
#> 50   1025 (1'000, 1'050]
#> 51   1025 (1'000, 1'050]
#> 52   1025 (1'000, 1'050]
#> 53   1025 (1'000, 1'050]
#> 54   1025 (1'000, 1'050]
#> 55   1025 (1'000, 1'050]
#> 56   1025 (1'000, 1'050]
#> 57   1025 (1'000, 1'050]
#> 58   1025 (1'000, 1'050]
#> 59   1025 (1'000, 1'050]
#> 60   1025 (1'000, 1'050]
#> 61   1025 (1'000, 1'050]
#> 62   1075 (1'050, 1'100]
#> 63   1075 (1'050, 1'100]
#> 64   1075 (1'050, 1'100]
#> 65   1075 (1'050, 1'100]
#> 66   1075 (1'050, 1'100]
#> 67   1075 (1'050, 1'100]
#> 68   1075 (1'050, 1'100]
#> 69   1075 (1'050, 1'100]
#> 70   1075 (1'050, 1'100]
#> 71   1075 (1'050, 1'100]
#> 72   1075 (1'050, 1'100]
#> 73   1075 (1'050, 1'100]
#> 74   1075 (1'050, 1'100]
#> 75   1075 (1'050, 1'100]
#> 76   1075 (1'050, 1'100]
#> 77   1075 (1'050, 1'100]
#> 78   1075 (1'050, 1'100]
#> 79   1075 (1'050, 1'100]
#> 80   1075 (1'050, 1'100]
#> 81   1075 (1'050, 1'100]
#> 82   1075 (1'050, 1'100]
#> 83   1075 (1'050, 1'100]
#> 84   1075 (1'050, 1'100]
#> 85   1075 (1'050, 1'100]
#> 86   1075 (1'050, 1'100]
#> 87   1075 (1'050, 1'100]
#> 88   1075 (1'050, 1'100]
#> 89   1075 (1'050, 1'100]
#> 90   1075 (1'050, 1'100]
#> 91   1075 (1'050, 1'100]
#> 92   1075 (1'050, 1'100]
#> 93   1075 (1'050, 1'100]
#> 94   1075 (1'050, 1'100]
#> 95   1075 (1'050, 1'100]
#> 96   1075 (1'050, 1'100]
#> 97   1075 (1'050, 1'100]
#> 98   1075 (1'050, 1'100]
#> 99   1075 (1'050, 1'100]
#> 100  1075 (1'050, 1'100]
#> 101  1075 (1'050, 1'100]
#> 102  1075 (1'050, 1'100]
#> 103  1075 (1'050, 1'100]
#> 104  1075 (1'050, 1'100]
#> 105  1075 (1'050, 1'100]
#> 106  1075 (1'050, 1'100]
#> 107  1075 (1'050, 1'100]
#> 108  1075 (1'050, 1'100]
#> 109  1075 (1'050, 1'100]
#> 110  1075 (1'050, 1'100]
#> 111  1075 (1'050, 1'100]
#> 
#> $breaks
#> [1]  950 1000 1050 1100
#> 
#> $bin_means
#> [1]  975 1025 1075
#> 
#> $bin_labels
#> [1] [950, 1'000]   (1'000, 1'050] (1'050, 1'100]
#> Levels: [950, 1'000] (1'000, 1'050] (1'050, 1'100]
#> 

auto_cut(c(0.0001, 0.0002, 0.0003, 0.005), n_bins = 3, format = "fg")

#> $data
#>   value          level
#> 1 0.001     [0, 0.002]
#> 2 0.001     [0, 0.002]
#> 3 0.001     [0, 0.002]
#> 4 0.005 (0.004, 0.006]
#> 
#> $breaks
#> [1] 0.000 0.002 0.004 0.006
#> 
#> $bin_means
#> [1] 0.001 0.003 0.005
#> 
#> $bin_labels
#> [1] [0, 0.002]     (0.002, 0.004] (0.004, 0.006]
#> Levels: [0, 0.002] (0.002, 0.004] (0.004, 0.006]
#>