Grouped Weighted Means, Quartiles, or Variances

Calculates weighted means, quartiles, or variances (and counts) of a variable grouped by optional columns. By default, counts are not weighted, even if there is a weighting variable.

grouped_stats(
  data,
  x,
  w = NULL,
  by = NULL,
  stats = c("mean", "quartiles", "variance"),
  counts = TRUE,
  counts_weighted = FALSE,
  counts_name = "counts",
  value_name = x,
  q1_name = "q1",
  q3_name = "q3",
  ...
)

Arguments

data

A data.frame.

x

Variable name in data to summarize.

w

Optional name of the column in data with case weights.

by

An optional vector of column names in data used to group the results.

stats

Statistic to calculate: "mean", "quartiles", or "variance".

counts

Should group counts be added?

counts_weighted

Should counts be weighted by the case weights? If TRUE, the sum of w is returned by group.

counts_name

Name of column in the resulting data.frame containing the counts.

value_name

Name of the resulting column with mean, median, or variance.

q1_name

Name of the resulting column with first quartile values. Only relevant if stats = "quartiles".

q3_name

Name of the resulting column with third quartile values. Only relevant if stats = "quartiles".

...

Additional arguments passed to corresponding weighted_*() functions in MetricsWeighted.

Value

A data.frame with columns by, x, and optionally counts_name.

Examples

grouped_stats(iris, "Sepal.Width")
#>   counts Sepal.Width
#> 1    150    3.057333
grouped_stats(iris, "Sepal.Width", stats = "quartiles")
#>   counts  q1 Sepal.Width  q3
#> 1    150 2.8           3 3.3
grouped_stats(iris, "Sepal.Width", stats = "variance")
#>   counts Sepal.Width
#> 1    150   0.1899794
grouped_stats(iris, "Sepal.Width", w = "Petal.Width", counts_weighted = TRUE)
#>   counts Sepal.Width
#> 1  179.9    2.956587
grouped_stats(iris, "Sepal.Width", by = "Species")
#>      Species counts Sepal.Width
#> 1     setosa     50       3.428
#> 2 versicolor     50       2.770
#> 3  virginica     50       2.974