Partial Dependence and other Profiles

Calculates different types of profiles across covariable values. By default, partial dependence profiles are calculated (see Friedman). Other options are profiles of ALE (accumulated local effects, see Apley), response, predicted values ("M plots" or "marginal plots", see Apley), residuals, and shap. The results are aggregated either by (weighted) means or by (weighted) quartiles. Note that ALE profiles are calibrated by (weighted) average predictions. In contrast to the suggestions in Apley, we calculate ALE profiles of factors in the same order as the factor levels. They are not being reordered based on similiarity of other variables.

light_profile(x, ...)

# S3 method for default
light_profile(x, ...)

# S3 method for flashlight
light_profile(
  x,
  v = NULL,
  data = NULL,
  by = x$by,
  type = c("partial dependence", "ale", "predicted", "response", "residual", "shap"),
  stats = c("mean", "quartiles"),
  breaks = NULL,
  n_bins = 11,
  cut_type = c("equal", "quantile"),
  use_linkinv = TRUE,
  value_name = "value",
  q1_name = "q1",
  q3_name = "q3",
  label_name = "label",
  type_name = "type",
  counts_name = "counts",
  counts = TRUE,
  counts_weighted = FALSE,
  v_labels = TRUE,
  pred = NULL,
  pd_evaluate_at = NULL,
  pd_grid = NULL,
  pd_indices = NULL,
  pd_n_max = 1000,
  pd_seed = NULL,
  pd_center = c("no", "first", "middle", "last", "mean", "0"),
  ale_two_sided = FALSE,
  ...
)

# S3 method for multiflashlight
light_profile(
  x,
  v = NULL,
  data = NULL,
  breaks = NULL,
  n_bins = 11,
  cut_type = c("equal", "quantile"),
  pd_evaluate_at = NULL,
  pd_grid = NULL,
  ...
)

Arguments

x
An object of class flashlight or multiflashlight.

...
Further arguments passed to cut3 resp. formatC in forming the cut breaks of the v variable. Not relevant for partial dependence and ALE profiles.

v
The variable to be profiled.

data
An optional data.frame. Not used for type = "shap".

by
An optional vector of column names used to additionally group the results.

type
Type of the profile: Either "partial dependence", "ale", "predicted", "response", "residual", or "shap".

stats
Statistic to calculate: "mean" or "quartiles". For ALE profiles, only "mean" makes sense.

breaks
Cut breaks for a numeric v.

n_bins
Maxmium number of unique values to evaluate for numeric v. Only used if neither grid nor pd_evaluate_at is specified.

cut_type
For the default "equal", bins of equal width are created for v by pretty. Choose "quantile" to create quantile bins.

use_linkinv
Should retransformation function be applied? Default is TRUE. Not used for type "shap".

value_name
Column name in resulting data containing the profile value. Defaults to "value".

q1_name
Name of the resulting column with first quartile values. Only relevant for stats "quartiles".

q3_name
Name of the resulting column with third quartile values. Only relevant for stats "quartiles".

label_name
Column name in resulting data containing the label of the flashlight. Defaults to "label".

type_name
Column name in the resulting data with the plot type.

counts_name
Name of the column containing counts if counts is TRUE.

counts
Should counts be added?

counts_weighted
If counts is TRUE: Should counts be weighted by the case weights? If TRUE, the sum of w is returned by group.

v_labels
If FALSE, return group centers of v instead of labels. Only relevant for types "response", "predicted" or "residual" and if v is being binned. In that case useful if e.g. different flashlights use different data sets and bin labels would not match.

pred
Optional vector with predictions (after application of inverse link). Can be used to avoid recalculation of predictions over and over if the functions is to be repeatedly called for different v and predictions are computationally expensive to make. Only relevant for type = "predicted" and type = "ale".

pd_evaluate_at
Vector with values of v used to evaluate the profile. Only relevant for type = "partial dependence" and "ale".

pd_grid
A data.frame with grid values, e.g. generated by expand.grid. Only used for type = "partial dependence".

pd_indices
A vector of row numbers to consider in calculating partial dependence profiles. Only used for type = "partial dependence" and "ale".

pd_n_max
Maximum number of ICE profiles to calculate (will be randomly picked from data). Only used for type = "partial dependence" and "ale".

pd_seed
Integer random seed used to select ICE profiles. Only used for type = "partial dependence" and "ale".

pd_center
How should ICE curves be centered? Default is "no". Choose "first", "middle", or "last" to 0-center at specific evaluation points. Choose "mean" to center all profiles at the within-group means. Choose "0" to mean-center curves at 0. Only relevant for partial dependence.

ale_two_sided
If TRUE, v is continuous and breaks are passed or being calculated, then two-sided derivatives are calculated for ALE instead of left derivatives. More specifically: Usually, local effects at value x are calculated using points between x-e and x. Set ale_two_sided = TRUE to use points between x-e/2 and x+e/2.

Value

An object of classes light_profile, light (and a list) with the following elements.

data A tibble containing results. Can be used to build fully customized visualizations. Its column names are specified by all other items in this list.

by Names of group by variable.

v The variable(s) evaluated.

type Same as input type. For information only.

stats Same as input stats.

value_name Same as input value_name.

q1_name Same as input q1_name.

q3_name Same as input q3_name.

label_name Same as input label_name.

type_name Same as input type_name.

counts_name Same as input counts_name.

Details

For numeric covariables v with more than n_bins disjoint values, its values are binned. Alternatively, breaks can be provided to specify the binning. For partial dependence profiles (and partly also ALE profiles), this behaviour can be overritten either by providing a vector of evaluation points (pd_evaluate_at) or an evaluation pd_grid. By the latter we mean a data frame with column name(s) with a (multi-)variate evaluation grid. For partial dependence, ALE, and prediction profiles, "model", "predict_function", linkinv" and "data" are required. For response profiles its "y", "linkinv" and "data" and for shap profiles it is just "shap". "data" can be passed on the fly.

Methods (by class)

default: Default method not implemented yet.

flashlight: Profiles for flashlight.

multiflashlight: Profiles for multiflashlight.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Apley D. W. (2016). Visualizing the effects of predictor variables in black box supervised learning models.

See also

Examples

fit_full <- lm(Sepal.Length ~ ., data = iris)
mod_full <- flashlight(model = fit_full, label = "full", data = iris, y = "Sepal.Length")
light_profile(mod_full, v = "Species")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> data.frames (maximum 6 rows shown):
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Species    counts value label type              
#>   <fct>       <int> <dbl> <chr> <fct>             
#> 1 setosa        150  6.43 full  partial dependence
#> 2 versicolor    150  5.70 full  partial dependence
#> 3 virginica     150  5.40 full  partial dependence

light_profile(mod_full, v = "Species", type = "response")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> data.frames (maximum 6 rows shown):
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Species    counts value label type    
#>   <fct>       <int> <dbl> <chr> <fct>   
#> 1 setosa         50  5.01 full  response
#> 2 versicolor     50  5.94 full  response
#> 3 virginica      50  6.59 full  response

light_profile(mod_full, v = "Species", stats = "quartiles")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> data.frames (maximum 6 rows shown):
#> 
#>  data 
#> # A tibble: 3 x 7
#>   Species    counts    q1 value    q3 label type              
#>   <fct>       <int> <dbl> <dbl> <dbl> <chr> <fct>             
#> 1 setosa        150  5.14  6.71  7.26 full  partial dependence
#> 2 versicolor    150  4.42  5.98  6.54 full  partial dependence
#> 3 virginica     150  4.12  5.68  6.24 full  partial dependence

light_profile(mod_full, v = "Petal.Width", type = "residual")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> data.frames (maximum 6 rows shown):
#> 
#>  data 
#> # A tibble: 6 x 5
#>   Petal.Width counts    value label type    
#>   <fct>        <int>    <dbl> <chr> <fct>   
#> 1 [0, 0.2]        34 -0.0274  full  residual
#> 2 (0.2, 0.4]      14  0.0669  full  residual
#> 3 (0.4, 0.6]       2 -0.00221 full  residual
#> 4 (0.8, 1]         7  0.0968  full  residual
#> 5 (1, 1.2]         8 -0.0584  full  residual
#> 6 (1.2, 1.4]      21  0.0383  full  residual

light_profile(mod_full, v = "Petal.Width", pd_evaluate_at = 2:4)

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> data.frames (maximum 6 rows shown):
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Petal.Width counts value label type              
#>         <int>  <int> <dbl> <chr> <fct>             
#> 1           2    150  5.59 full  partial dependence
#> 2           3    150  5.28 full  partial dependence
#> 3           4    150  4.96 full  partial dependence