Most Important Variables.

Returns the most important variable names sorted descendingly.

most_important(x, top_m = Inf)

# S3 method for default
most_important(x, top_m = Inf)

# S3 method for light_importance
most_important(x, top_m = Inf)

Arguments

x
An object of class light_importance.

top_m
Maximum number of important variables to be returned. Defaults to Inf, i.e. return all variables in descending order of importance.

Value

A character vector of variable names sorted in descending order by importance.

Methods (by class)

default: Default method not implemented yet.

light_importance: Extracts most important variables from an object of class light_importance.

See also

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
fl <- flashlight(model = fit, label = "ols", data = iris, y = "Sepal.Length")
(imp <- light_importance(fl, seed = 4))

#> 
#> I am an object with class(es) light_importance, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 4 x 6
#>   variable     metric value_shuffled label value_original value
#>   <chr>        <fct>           <dbl> <chr>          <dbl> <dbl>
#> 1 Sepal.Width  rmse            0.403 ols            0.301 0.102
#> 2 Petal.Length rmse            1.96  ols            0.301 1.66 
#> # ... with 2 more rows

most_important(imp)

#> [1] "Petal.Length" "Species"      "Petal.Width"  "Sepal.Width" 

most_important(imp, 2)

#> [1] "Petal.Length" "Species"