Individual Conditional Expectation (ICE)

Generates Individual Conditional Expectation (ICE) profiles. An ICE profile shows how the prediction of an observation changes if one or multiple variables are systematically changed across its ranges, holding all other values fixed [1]. The curves can be centered in order to increase visibility of interaction effects.

light_ice(x, ...)

# S3 method for default
light_ice(x, ...)

# S3 method for flashlight
light_ice(
  x,
  v = NULL,
  data = x$data,
  by = x$by,
  evaluate_at = NULL,
  breaks = NULL,
  grid = NULL,
  n_bins = 27,
  cut_type = c("equal", "quantile"),
  indices = NULL,
  n_max = 20,
  seed = NULL,
  use_linkinv = TRUE,
  center = c("no", "first", "middle", "last", "mean", "0"),
  value_name = "value",
  label_name = "label",
  id_name = "id",
  ...
)

# S3 method for multiflashlight
light_ice(x, ...)

Arguments

x
An object of class flashlight or multiflashlight.

...
Further arguments passed to or from other methods.

v
The variable to be profiled.

data
An optional data.frame.

by
An optional vector of column names used to additionally group the results.

evaluate_at
Vector with values of v used to evaluate the profile.

breaks
Instead of evaluate_at (and grid), cut points for x can be provided. From them, evaluate_at values are calculates as averages.

grid
A data.frame with grid values as those generated by expand.grid.

n_bins
Maximum number of unique values to evaluate for numeric v. Only used in neither grid nor evaluate_at is specified.

cut_type
For the default "equal", bins of equal width are created for v by pretty. Choose "quantile" to create quantile bins. Only used in neither grid nor evaluate_at is specified.

indices
A vector of row numbers to consider.

n_max
If indices is not given, maximum number of rows to consider. Will be randomly picked from data if necessary.

seed
An integer random seed.

use_linkinv
Should retransformation function be applied? Default is TRUE.

center
How should curves be centered? Default is "no". Choose "first", "middle", or "last" to 0-center at specific evaluation points. Choose "mean" to center all profiles at the within-group means. Choose "0" to mean-center curves at 0.

value_name
Column name in resulting data containing the profile value. Defaults to "value".

label_name
Column name in resulting data containing the label of the flashlight. Defaults to "label".

id_name
Column name in resulting data containing the row id of the profile. Defaults to "id_name".

Value

An object of class light_ice, light (and a list) with the following elements.

data A tibble containing the results. Can be used to build fully customized visualizations. Its column names are specified by all other items in this list.

by Same as input by.

v The variable(s) evaluated.

center How centering was done.

value_name Same as input value_name.

label_name Same as input label_name.

id_name Same as input id_name.

Details

There are two ways to specify the variable(s) to be profiled. The first option is to pass the variable name via v and an optional vector with evaluation points evaluate_at (or breaks). This works for dependence on a single variable. The second option is much more general: You can specify any grid as a data.frame with one or more columns. It can e.g. be generated by a call to expand.grid. Currently, there is no option to pass more than one variable name without such grid. The minimum required elements in the (multi-)flashlight are "predict_function", "model", "linkinv" and "data", where the latest can be passed on the fly. Which rows in data are profiled? This is specified by indices. If not given and n_max is smaller than the number of rows in data, then row indices will be sampled randomly from data. If the same rows should be used for all flashlights in a multiflashlight, there are two options: Either pass a seed (with potentially undesired consequences for subsequent code) or a vector of indices used to select rows. In both cases, data should be the same for all flashlights considered.

Methods (by class)

default: Default method not implemented yet.

flashlight: ICE profiles for a flashlight object.

multiflashlight: ICE profiles for a multiflashlight object.

References

[1] Goldstein, A. et al. (2015). Peeking inside the black box: Visualizing statistical learning with plots of individual conditional expectation. Journal of Computational and Graphical Statistics, 24:1 <doi.org/10.1080/10618600.2014.907095>.

See also

Examples

fit_full <- lm(Sepal.Length ~ ., data = iris)
fit_part <- lm(Sepal.Length ~ Petal.Length, data = iris)
mod_full <- flashlight(model = fit_full, label = "full", data = iris, y = "Sepal.Length")
mod_part <- flashlight(model = fit_part, label = "part", data = iris, y = "Sepal.Length")
mods <- multiflashlight(list(mod_full, mod_part))
grid <- expand.grid(Species = levels(iris$Species), Petal.Length = 2:4)
light_ice(mod_full, v = "Species")

#> 
#> I am an object with class(es) light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   id    Species    value label
#> 1 10     setosa 4.920872  full
#> 2 10 versicolor 4.197310  full
#> 3 10  virginica 3.897375  full
#> 4 74     setosa 7.079015  full
#> 5 74 versicolor 6.355454  full
#> 6 74  virginica 6.055518  full

light_ice(mod_full, v = "Species", indices = (1:15) * 10)

#> 
#> I am an object with class(es) light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   id    Species    value label
#> 1 10     setosa 4.920872  full
#> 2 10 versicolor 4.197310  full
#> 3 10  virginica 3.897375  full
#> 4 20     setosa 5.204964  full
#> 5 20 versicolor 4.481402  full
#> 6 20  virginica 4.181466  full

light_ice(mod_full, v = "Species", evaluate_at = levels(iris$Species))

#> 
#> I am an object with class(es) light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   id    Species    value label
#> 1 35     setosa 4.889357  full
#> 2 35 versicolor 4.165795  full
#> 3 35  virginica 3.865859  full
#> 4 16     setosa 5.470981  full
#> 5 16 versicolor 4.747419  full
#> 6 16  virginica 4.447484  full

light_ice(mod_full, grid = grid, data = iris[1,])$data

#> # A tibble: 9 x 5
#>      id Species    Petal.Length value label
#>   <int> <fct>             <int> <dbl> <chr>
#> 1     1 setosa                2  5.50 full 
#> 2     1 versicolor            2  4.78 full 
#> 3     1 virginica             2  4.48 full 
#> 4     1 setosa                3  6.33 full 
#> 5     1 versicolor            3  5.61 full 
#> 6     1 virginica             3  5.31 full 
#> 7     1 setosa                4  7.16 full 
#> 8     1 versicolor            4  6.44 full 
#> 9     1 virginica             4  6.14 full 

light_ice(mods, v = "Species", indices = (1:15) * 10)

#> 
#> I am an object with class(es) light_ice_multi, light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   id    Species    value label
#> 1 10     setosa 4.920872  full
#> 2 10 versicolor 4.197310  full
#> 3 10  virginica 3.897375  full
#> 4 20     setosa 5.204964  full
#> 5 20 versicolor 4.481402  full
#> 6 20  virginica 4.181466  full

light_ice(mods, v = "Species", indices = (1:15) * 10, center = "first")

#> 
#> I am an object with class(es) light_ice_multi, light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   id    Species     value label
#> 1 10     setosa  0.000000  full
#> 2 10 versicolor -0.723562  full
#> 3 10  virginica -1.023498  full
#> 4 20     setosa  0.000000  full
#> 5 20 versicolor -0.723562  full
#> 6 20  virginica -1.023498  full

light_ice(mods, v = "Petal.Width", n_bins = 5)

#> 
#> I am an object with class(es) light_ice_multi, light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   id Petal.Width    value label
#> 1 36        0.25 4.674415  full
#> 2 36        0.75 4.516837  full
#> 3 36        1.25 4.359260  full
#> 4 36        1.75 4.201682  full
#> 5 36        2.25 4.044104  full
#> 6 99        0.25 5.096370  full

light_ice(mods, v = "Petal.Width", by = "Species", n_bins = 5)

#> 
#> I am an object with class(es) light_ice_multi, light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   id Species Petal.Width    value label
#> 1 20  setosa        0.25 5.220721  full
#> 2 20  setosa        0.75 5.063144  full
#> 3 20  setosa        1.25 4.905566  full
#> 4 20  setosa        1.75 4.747989  full
#> 5 20  setosa        2.25 4.590411  full
#> 6 19  setosa        0.25 5.386570  full

light_ice(mods, v = "Petal.Width", by = "Species",
  id_name = "profile", value_name = "val", label_name = "lab")

#> 
#> I am an object with class(es) light_ice_multi, light_ice, light, list 
#> 
#> data.frames (head):
#> 
#>  data 
#>   profile   Species Petal.Width      val  lab
#> 1     125 virginica         0.1 7.479377 full
#> 2     125 virginica         0.2 7.447861 full
#> 3     125 virginica         0.3 7.416346 full
#> 4     125 virginica         0.4 7.384830 full
#> 5     125 virginica         0.5 7.353315 full
#> 6     125 virginica         0.6 7.321799 full