Permutation Importance

Calculates performance per variable with respect to a performance measure before and after permuting its values. The difference is a measure of importance, see Fisher et al. 2018 [1]. Using more than one m_repetitions resp. permutation will allow more accurate assessment of variable importance as well as the calculation of standard errors.

light_importance(x, ...)

# S3 method for default
light_importance(x, ...)

# S3 method for flashlight
light_importance(
  x,
  data = x$data,
  by = x$by,
  metric = x$metrics[1],
  v = NULL,
  n_max = Inf,
  seed = NULL,
  m_repetitions = 1,
  lower_is_better = TRUE,
  use_linkinv = FALSE,
  metric_name = "metric",
  value_name = "value",
  error_name = "error",
  label_name = "label",
  variable_name = "variable",
  ...
)

# S3 method for multiflashlight
light_importance(x, ...)

Arguments

x
An object of class flashlight or multiflashlight.

...
Further arguments passed to light_performance.

data
An optional data.frame.

by
An optional vector of column names used to additionally group the results.

metric
An optional named list of length one with a metric as element. Defaults to the first metric in the flashlight. The metric needs to be a function with at least four arguments: actual, predicted, case weights w and ....

v
Vector of variables to assess importance for. Defaults to all variables in data.

n_max
Maximum number of rows to consider. Use if data is large.

seed
An integer random seed used to select and shuffle rows.

m_repetitions
Number of permutations. Defaults to 1. A value above 1 provides more stable estimates of variable importance along with the calculation of standard errors.

lower_is_better
Logical flag indicating if lower values in the metric are better or not. If set to FALSE, the increase in metric is multiplied by -1.

use_linkinv
Should retransformation function be applied? Default is FALSE.

metric_name
Name of the resulting column containing the name of the metric. Defaults to "metric".

value_name
Column name in resulting data containing the (average) drop in performance. Defaults to "value".

error_name
Column name in resulting data containing the standard error of drop in performance. Defaults to "error". NA if m_repetitions = 1.

label_name
Column name in resulting data containing the label of the flashlight. Defaults to "label".

variable_name
Column name in resulting data containing the variable names. Defaults to "variable".

Value

An object of class light_importance, light (and a list) with the following elements.

data A tibble with results. Can be used to build fully customized visualizations. The columns "value_original" and "value_shuffled" provide the performance before and after shuffling.

by Same as input by.

metric_name Column name representing the name of the metric. For information only.

value_name Same as input value_name.

error_name Same as input error_name.

label_name Same as input label_name.

variable_name Same as input variable_name.

Details

The minimum required elements in the (multi-) flashlight are "y", "predict_function", "model", "data" and "metrics". The latter two can also directly be passed to light_importance. Note that by default, no retransformation function is applied.

Methods (by class)

default: Default method not implemented yet.

flashlight: Variable importance for a flashlight.

multiflashlight: Variable importance for a multiflashlight.

References

[1] Fisher A., Rudin C., Dominici F. (2018). All Models are Wrong but many are Useful: Variable Importance for Black-Box, Proprietary, or Misspecified Prediction Models, using Model Class Reliance. ArXiv <arxiv.org/abs/1801.01489>.

See also

Examples

fit_part <- lm(Sepal.Length ~ Petal.Length, data = iris)
fit_full <- lm(Sepal.Length ~ ., data = iris)
mod_full <- flashlight(model = fit_full, label = "full", data = iris, y = "Sepal.Length")
mod_part <- flashlight(model = fit_part, label = "part", data = iris, y = "Sepal.Length")
mods <- multiflashlight(list(mod_full, mod_part), by = "Species")
light_importance(mod_full)

#> 
#> I am an object with class(es) light_importance, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 4 x 7
#>   label metric variable     value_shuffled value_original  value error
#>   <chr> <fct>  <chr>                 <dbl>          <dbl>  <dbl> <lgl>
#> 1 full  rmse   Sepal.Width           0.390          0.301 0.0898 NA   
#> 2 full  rmse   Petal.Length          2.02           0.301 1.72   NA   
#> # ... with 2 more rows