Partial Dependence and other Profiles

Calculates different types of profiles across covariable values. By default, partial dependence profiles are calculated (see Friedman). Other options are profiles of ALE (accumulated local effects, see Apley), response, predicted values ("M plots" or "marginal plots", see Apley), residuals, and shap. The results are aggregated either by (weighted) means or by (weighted) quartiles.

Note that ALE profiles are calibrated by (weighted) average predictions. In contrast to the suggestions in Apley, we calculate ALE profiles of factors in the same order as the factor levels. They are not being reordered based on similiarity of other variables.

light_profile(x, ...)

# S3 method for default
light_profile(x, ...)

# S3 method for flashlight
light_profile(
  x,
  v = NULL,
  data = NULL,
  by = x$by,
  type = c("partial dependence", "ale", "predicted", "response", "residual", "shap"),
  stats = c("mean", "quartiles"),
  breaks = NULL,
  n_bins = 11L,
  cut_type = c("equal", "quantile"),
  use_linkinv = TRUE,
  counts = TRUE,
  counts_weighted = FALSE,
  v_labels = TRUE,
  pred = NULL,
  pd_evaluate_at = NULL,
  pd_grid = NULL,
  pd_indices = NULL,
  pd_n_max = 1000L,
  pd_seed = NULL,
  pd_center = c("no", "first", "middle", "last", "mean", "0"),
  ale_two_sided = FALSE,
  ...
)

# S3 method for multiflashlight
light_profile(
  x,
  v = NULL,
  data = NULL,
  type = c("partial dependence", "ale", "predicted", "response", "residual", "shap"),
  breaks = NULL,
  n_bins = 11L,
  cut_type = c("equal", "quantile"),
  pd_evaluate_at = NULL,
  pd_grid = NULL,
  ...
)

Arguments

x

An object of class "flashlight" or "multiflashlight".

...

Further arguments passed to cut3() in forming the cut breaks of the v variable.

v

The variable name to be profiled.

data

An optional data.frame. Not used for type = "shap".

by

An optional vector of column names used to additionally group the results.

type

Type of the profile: Either "partial dependence", "ale", "predicted", "response", "residual", or "shap".

stats

Statistic to calculate: "mean" or "quartiles". For ALE profiles, only "mean" makes sense.

breaks

Cut breaks for a numeric v. Used to overwrite automatic binning via n_bins and cut_type. Ignored if v is not numeric.

n_bins

Approximate number of unique values to evaluate for numeric v. Ignored if v is not numeric or if breaks is specified.

cut_type

Should a numeric v be cut into "equal" or "quantile" bins? Ignored if v is not numeric or if breaks is specified.

use_linkinv

Should retransformation function be applied? Default is TRUE. Not used for type "shap".

counts

Should observation counts be added?

counts_weighted

If counts = TRUE: Should counts be weighted by the case weights? If TRUE, the sum of w is returned by group.

v_labels

If FALSE, return group centers of v instead of labels. Only relevant for types "response", "predicted" or "residual" and if v is being binned. In that case useful, for instance, if different flashlights use different data sets and bin labels would not match.

pred

Optional vector with predictions (after application of inverse link). Can be used to avoid recalculation of predictions over and over if the functions is to be repeatedly called for different v and predictions are computationally expensive to make. Not implemented for multiflashlight.

pd_evaluate_at

Vector with values of v used to evaluate the profile. Only relevant for type = "partial dependence" and "ale".

pd_grid

A data.frame with grid values, e.g., generated by expand.grid(). Only used for type = "partial dependence".

pd_indices

A vector of row numbers to consider in calculating partial dependence profiles and "ale".

pd_n_max

Maximum number of ICE profiles to calculate (will be randomly picked from data) for partial dependence and ALE.

pd_seed

Integer random seed used to select ICE profiles for partial dependence and ALE.

pd_center

How should ICE curves be centered?

Default is "no".

Choose "first", "middle", or "last" to 0-center at specific evaluation points.

Choose "mean" to center all profiles at the within-group means.

Choose "0" to mean-center curves at 0. Only relevant for partial dependence.

ale_two_sided

If TRUE, v is continuous and breaks are passed or being calculated, then two-sided derivatives are calculated for ALE instead of left derivatives. More specifically: Usually, local effects at value x are calculated using points in \([x-e, x]\). Set ale_two_sided = TRUE to use points in \([x-e/2, x+e/2]\).

Value

An object of class "light_profile" with the following elements:

data A tibble containing results. Can be used to build fully customized visualizations. Column names can be controlled by options(flashlight.column_name).

by Names of group by variable.

v The variable(s) evaluated.

type Same as input type. For information only.

stats Same as input stats.

Details

Numeric covariables v with more than n_bins disjoint values are binned into n_bins bins. Alternatively, breaks can be provided to specify the binning. For partial dependence profiles (and partly also ALE profiles), this behaviour can be overwritten either by providing a vector of evaluation points (pd_evaluate_at) or an evaluation pd_grid. By the latter we mean a data frame with column name(s) with a (multi-)variate evaluation grid.

For partial dependence, ALE, and prediction profiles, "model", "predict_function", "linkinv" and "data" are required. For response profiles its "y", "linkinv" and "data", and for shap profiles it is just "shap". "data" can be passed on the fly.

Methods (by class)

light_profile(default): Default method not implemented yet.

light_profile(flashlight): Profiles for flashlight.

light_profile(multiflashlight): Profiles for multiflashlight.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Apley D. W. (2016). Visualizing the effects of predictor variables in black box supervised learning models.

See also

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
fl <- flashlight(model = fit, label = "iris", data = iris, y = "Sepal.Length")
light_profile(fl, v = "Species")
#> 
#> I am an object of class light_profile 
#> 
#> data.frames:
#> 
#>  data 
#>      Species counts    value label               type
#> 1     setosa    150 6.425687  iris partial dependence
#> 2 versicolor    150 5.702125  iris partial dependence
#> 3  virginica    150 5.402189  iris partial dependence
light_profile(fl, v = "Petal.Width", type = "residual")
#> 
#> I am an object of class light_profile 
#> 
#> data.frames:
#> 
#>  data 
#>    Petal.Width counts        value label     type
#> 1     [0, 0.2]     34 -0.027433217  iris residual
#> 2   (0.2, 0.4]     14  0.066938641  iris residual
#> 3   (0.4, 0.6]      2 -0.002205793  iris residual
#> 4     (0.8, 1]      7  0.096800589  iris residual
#> 5     (1, 1.2]      8 -0.058353620  iris residual
#> 6   (1.2, 1.4]     21  0.038269727  iris residual
#> 7   (1.4, 1.6]     16 -0.050392615  iris residual
#> 8   (1.6, 1.8]     14 -0.112246236  iris residual
#> 9     (1.8, 2]     11  0.041651915  iris residual
#> 10    (2, 2.2]      9  0.027919108  iris residual
#> 11  (2.2, 2.4]     11  0.122917919  iris residual
#> 12  (2.4, 2.6]      3 -0.232750147  iris residual