Partial Dependence and other Profiles

Calculates different types of profiles across covariable values. By default, partial dependence profiles are calculated but also profiles of response, predicted values and residuals are possible. The results are aggregated either by (weighted) means or by (weighted) quartiles. Furthermore, counts are added to the resulting object.

light_profile(x, ...)

# S3 method for default
light_profile(x, ...)

# S3 method for flashlight
light_profile(x, v = NULL, data = NULL,
  by = x$by, type = c("partial dependence", "predicted", "response",
  "residual"), stats = c("mean", "quartiles"), breaks = NULL,
  n_bins = 11, cut_type = c("equal", "quantile"), use_linkinv = TRUE,
  value_name = "value", q1_name = "q1", q3_name = "q3",
  label_name = "label", type_name = "type", counts_name = "counts",
  counts = TRUE, counts_weighted = FALSE, v_labels = TRUE,
  pred = NULL, pd_evaluate_at = NULL, pd_grid = NULL,
  pd_indices = NULL, pd_n_max = 1000, pd_seed = NULL, ...)

# S3 method for multiflashlight
light_profile(x, v = NULL, data = NULL,
  breaks = NULL, n_bins = 11, cut_type = c("equal", "quantile"),
  pd_evaluate_at = NULL, pd_grid = NULL, ...)

Arguments

x
An object of class flashlight or multiflashlight.

...
Further arguments passed to cut3 resp. formatC in forming the cut breaks of the v variable. Not relevant for partial dependence profiles.

v
The variable to be profiled.

data
An optional data.frame.

by
An optional vector of column names used to additionally group the results.

type
Type of the profile: Either "partial dependence", predicted", "response", or "residual".

stats
Statistic to calculate: "mean" or "quartiles".

breaks
Cut breaks for a numeric v.

n_bins
Maxmium number of unique values to evaluate for numeric v. Only used in neither grid nor evaluate_at is specified.

cut_type
For the default "equal", bins of equal width are created for v by pretty. Choose "quantile" to create quantile bins.

use_linkinv
Should retransformation function be applied? Default is TRUE.

value_name
Column name in resulting data containing the profile value. Defaults to "value".

q1_name
Name of the resulting column with first quartile values. Only relevant for stats "quartiles".

q3_name
Name of the resulting column with third quartile values. Only relevant for stats "quartiles".

label_name
Column name in resulting data containing the label of the flashlight. Defaults to "label".

type_name
Column name in the resulting data with the plot type.

counts_name
Name of the column containing counts if counts is TRUE.

counts
Should counts be added?

counts_weighted
If counts is TRUE: Should counts be weighted by the case weights? If TRUE, the sum of w is returned by group.

v_labels
If FALSE, return group centers of v instead of labels. Only relevant if v is numeric with many distinct values. In that case useful if e.g. different flashlights use different data sets.

pred
Optional vector with predictions (after application of inverse link). Can be used to avoid recalculation of predictions over and over if the functions is to be repeatedly called for different v and predictions are computationally expensive to make. Only relevant for type = "predicted".

pd_evaluate_at
Vector with values of v used to evaluate the profile. Only relevant for type = "partial dependence".

pd_grid
A data.frame with grid values, e.g. generated by expand.grid. Only used for type = "partial dependence".

pd_indices
A vector of row numbers to consider in calculating partial dependence profiles. Only used for type = "partial dependence".

pd_n_max
Maximum number of ICE profiles to calculate (will be randomly picked from data). Only used if type = "partial dependence".

pd_seed
Integer random seed used to select ICE profiles. Only used for type = "partial dependence".

Value

An object of classes light_profile, light (and a list) with the following elements.

data A tibble containing results. Can be used to build fully customized visualizations. Its column names are specified by all other items in this list.

by Names of group by variable.

v The variable(s) evaluated.

type Same as input type. For information only.

stats Same as input stats.

value_name Same as input value_name.

q1_name Same as input q1_name.

q3_name Same as input q3_name.

label_name Same as input label_name.

type_name Same as input type_name.

counts_name Same as input counts_name.

Details

For numeric covariables v with more than n_bins disjoint values, its values are binned. Alternatively, breaks can be provided to specify the binning. For partial dependence profiles, this behaviour can be overritten either by providing a vector of evaluation points (pd_evaluate_at) or an evaluation pd_grid. By the latter we mean a data frame with column name(s) with a (multi-)variate evaluation grid. For partial dependence or prediction profiles, "model", "predict_function", linkinv" and "data" are required. For response profiles its just "y", "linkinv" and "data". "data" can be passed on the fly for both types.

Methods (by class)

default: Default method not implemented yet.

flashlight: Profiles for flashlight.

multiflashlight: Profiles for multiflashlight.

See also

Examples

fit_full <- lm(Sepal.Length ~ ., data = iris)
fit_part <- lm(Sepal.Length ~ Petal.Length, data = iris)
mod_full <- flashlight(model = fit_full, label = "full", data = iris, y = "Sepal.Length")
mod_part <- flashlight(model = fit_part, label = "part", data = iris, y = "Sepal.Length")
mods <- multiflashlight(list(mod_full, mod_part))

light_profile(mod_full, v = "Species")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Species    counts value label type              
#>   <fct>       <int> <dbl> <chr> <fct>             
#> 1 setosa        150  6.43 full  partial dependence
#> 2 versicolor    150  5.70 full  partial dependence
#> # ... with 1 more row

light_profile(mod_full, v = "Species", counts = FALSE)

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 3 x 4
#>   Species    value label type              
#>   <fct>      <dbl> <chr> <fct>             
#> 1 setosa      6.43 full  partial dependence
#> 2 versicolor  5.70 full  partial dependence
#> # ... with 1 more row

light_profile(mod_full, v = "Species", type = "response")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Species    counts value label type    
#>   <fct>       <int> <dbl> <chr> <fct>   
#> 1 setosa         50  5.01 full  response
#> 2 versicolor     50  5.94 full  response
#> # ... with 1 more row

light_profile(mod_full, v = "Species", stats = "quartiles")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 3 x 7
#>   Species    counts    q1 value    q3 label type              
#>   <fct>       <int> <dbl> <dbl> <dbl> <chr> <fct>             
#> 1 setosa        150  5.14  6.71  7.26 full  partial dependence
#> 2 versicolor    150  4.42  5.98  6.54 full  partial dependence
#> # ... with 1 more row

light_profile(mod_full, v = "Petal.Width")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 13 x 5
#>   Petal.Width counts value label type              
#>         <dbl>  <int> <dbl> <chr> <fct>             
#> 1         0.1    150  6.19 full  partial dependence
#> 2         0.3    150  6.13 full  partial dependence
#> # ... with 11 more rows

light_profile(mod_full, v = "Petal.Width", type = "residual")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 12 x 5
#>   Petal.Width counts   value label type    
#>   <fct>        <int>   <dbl> <chr> <fct>   
#> 1 [0, 0.2]        34 -0.0274 full  residual
#> 2 (0.2, 0.4]      14  0.0669 full  residual
#> # ... with 10 more rows

light_profile(mod_full, v = "Petal.Width", type = "residual", v_label = FALSE)

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 12 x 5
#>   Petal.Width counts   value label type    
#>         <dbl>  <int>   <dbl> <chr> <fct>   
#> 1         0.1     34 -0.0274 full  residual
#> 2         0.3     14  0.0669 full  residual
#> # ... with 10 more rows

light_profile(mod_full, v = "Petal.Width", type = "residual", dig.lab = 1)

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 12 x 5
#>   Petal.Width counts   value label type    
#>   <fct>        <int>   <dbl> <chr> <fct>   
#> 1 [0, 0.2]        34 -0.0274 full  residual
#> 2 (0.2, 0.4]      14  0.0669 full  residual
#> # ... with 10 more rows

light_profile(mod_full, v = "Petal.Width", stats = "quartiles")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 13 x 7
#>   Petal.Width counts    q1 value    q3 label type              
#>         <dbl>  <int> <dbl> <dbl> <dbl> <chr> <fct>             
#> 1         0.1    150  5.21  6.35  6.92 full  partial dependence
#> 2         0.3    150  5.14  6.29  6.85 full  partial dependence
#> # ... with 11 more rows

light_profile(mod_full, v = "Petal.Width", n_bins = 3)

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Petal.Width counts value label type              
#>         <dbl>  <int> <dbl> <chr> <fct>             
#> 1         0.5    150  6.06 full  partial dependence
#> 2         1.5    150  5.75 full  partial dependence
#> # ... with 1 more row

light_profile(mod_full, v = "Petal.Width", pd_evaluate_at = 2:4)

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Petal.Width counts value label type              
#>         <int>  <int> <dbl> <chr> <fct>             
#> 1           2    150  5.59 full  partial dependence
#> 2           3    150  5.28 full  partial dependence
#> # ... with 1 more row

light_profile(mod_full, pd_grid = data.frame(Petal.Width = 2:4))

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 3 x 5
#>   Petal.Width counts value label type              
#>         <int>  <int> <dbl> <chr> <fct>             
#> 1           2    150  5.59 full  partial dependence
#> 2           3    150  5.28 full  partial dependence
#> # ... with 1 more row

light_profile(mod_full, v = "Petal.Width", by = "Species")

#> 
#> I am an object with class(es) light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 39 x 6
#>   Species Petal.Width counts value label type              
#>   <fct>         <dbl>  <int> <dbl> <chr> <fct>             
#> 1 setosa          0.1     50  5.05 full  partial dependence
#> 2 setosa          0.3     50  4.99 full  partial dependence
#> # ... with 37 more rows

light_profile(mods, v = "Petal.Width")

#> 
#> I am an object with class(es) light_profile_multi, light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 26 x 5
#>   Petal.Width counts value label type              
#>         <dbl>  <int> <dbl> <fct> <fct>             
#> 1         0.1    150  6.19 full  partial dependence
#> 2         0.3    150  6.13 full  partial dependence
#> # ... with 24 more rows

light_profile(mods, v = "Petal.Width", by = "Species")

#> 
#> I am an object with class(es) light_profile_multi, light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 78 x 6
#>   Species Petal.Width counts value label type              
#>   <fct>         <dbl>  <int> <dbl> <fct> <fct>             
#> 1 setosa          0.1     50  5.05 full  partial dependence
#> 2 setosa          0.3     50  4.99 full  partial dependence
#> # ... with 76 more rows

light_profile(mods, v = "Petal.Width", by = "Species", type = "predicted")

#> 
#> I am an object with class(es) light_profile_multi, light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 30 x 6
#>   Species Petal.Width counts value label type     
#>   <fct>   <fct>        <int> <dbl> <fct> <fct>    
#> 1 setosa  [0, 0.2]        34  4.98 full  predicted
#> 2 setosa  (0.2, 0.4]      14  5.07 full  predicted
#> # ... with 28 more rows

light_profile(mods, v = "Petal.Width", by = "Species",
  type = "predicted", stats = "quartiles")

#> 
#> I am an object with class(es) light_profile_multi, light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 30 x 8
#>   Species Petal.Width counts    q1 value    q3 label type     
#>   <fct>   <fct>        <int> <dbl> <dbl> <dbl> <fct> <fct>    
#> 1 setosa  [0, 0.2]        34  4.81  4.96  5.09 full  predicted
#> 2 setosa  (0.2, 0.4]      14  4.94  5.06  5.33 full  predicted
#> # ... with 28 more rows

light_profile(mods, v = "Petal.Width", by = "Species", stats = "quartiles",
  value_name = "pd", q1_name = "p25", q3_name = "p75", label_name = "model",
  type_name = "visualization", counts_name = "n")

#> 
#> I am an object with class(es) light_profile_multi, light_profile, light, list 
#> 
#> Tibbles:
#> 
#>  data 
#> # A tibble: 78 x 8
#>   Species Petal.Width     n   p25    pd   p75 model visualization     
#>   <fct>         <dbl> <int> <dbl> <dbl> <dbl> <fct> <fct>             
#> 1 setosa          0.1    50  4.92  5.06  5.21 full  partial dependence
#> 2 setosa          0.3    50  4.86  5.00  5.15 full  partial dependence
#> # ... with 76 more rows